Add an alternate contact to vendor schema
When dealing with vendors, it's helpful to have two points of
contact -- a main for current and routine communication, and an
alternate for when things go sideways (maybe main is on vacation;
maybe main lost their job, or died).

This PR adds an alternate contact to the vendor schema, allowing
for the inclusion of this additional contact information. Also it
adds a title to the contact, to illuminate potential
organizational relationships between the contacts -- for example,
is the alternate an assistant to the main, or a superior, or a
subordinate (which may be useful in guiding communication).